pyproject.toml
[project]
name = "synthds-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dask>=2025.3.0",
    "distributed>=2025.3.0",
    "kafka-python>=2.0.6",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "tqdm>=4.67.1",
]
